[#4] feat: 게시글에서 태그 모두 보이게 변경, 태그 크기 버그 해결
