+++ app/src/main/java/kr/cosine/groupfinder/presentation/view/common/decoration/AbstractTagItemDecoration.kt
@@ -1,0 +1,10 @@
+package kr.cosine.groupfinder.presentation.view.common.decoration
+
+import android.content.res.Resources
+import androidx.recyclerview.widget.RecyclerView
+
+abstract class AbstractTagItemDecoration : RecyclerView.ItemDecoration() {
+
+    protected val Int.dp: Int
+        get() = (this * Resources.getSystem().displayMetrics.density + 0.5f).toInt()
+}
