+++ app/src/main/java/kr/cosine/groupfinder/presentation/view/detail/DetailActivity.kt
@@ -28,9 +28,10 @@ import kr.cosine.groupfinder.presentation.view.common.data.IntentKey
 import kr.cosine.groupfinder.presentation.view.common.extension.applyWhite
 import kr.cosine.groupfinder.presentation.view.common.extension.setOnClickListenerWithCooldown
 import kr.cosine.groupfinder.presentation.view.common.extension.showToast
+import kr.cosine.groupfinder.presentation.view.common.flexbox.decoration.FlexboxItemDecoration
+import kr.cosine.groupfinder.presentation.view.common.flexbox.manager.FlexboxLayoutManager
 import kr.cosine.groupfinder.presentation.view.detail.event.DetailEvent
 import kr.cosine.groupfinder.presentation.view.dialog.Dialog
-import kr.cosine.groupfinder.presentation.view.group.adapter.decoration.GroupTagItemDecoration
 import kr.cosine.groupfinder.util.MyFirebaseMessagingService
 import java.util.UUID
 
@@ -120,9 +121,10 @@ class DetailActivity : GroupFinderActivity() {
     private fun bindCategoriesFromData(tags: List<String>) {
         with(binding.tags) {
             adapter = categoryAdapter
-            layoutManager = LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
-            removeItemDecoration(GroupTagItemDecoration)
-            addItemDecoration(GroupTagItemDecoration)
+            layoutManager = FlexboxLayoutManager(context)
+            if (itemDecorationCount == 0) {
+                addItemDecoration(FlexboxItemDecoration)
+            }
         }
         categoryAdapter.categoriesUpdate(tags)
     }
