--- app/src/main/res/drawable/flexbox_grid_tag_decoration.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <corners android:radius="0dp" />
-    <size
-        android:width="6dp"
-        android:height="6dp" />
-    <solid android:color="@android:color/transparent" />
-</shape>
