+++ app/src/main/res/layout/activity_detail.xml
@@ -57,11 +57,11 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="13dp"
             android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            android:overScrollMode="never"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/idTextView"
-            tools:itemCount="6"
+            tools:itemCount="1"
             tools:listitem="@layout/item_tag" />
 
         <ImageButton
