+++ app/src/main/res/layout/activity_write.xml
@@ -162,6 +162,7 @@
                 android:layout_marginTop="8dp"
                 android:layout_marginEnd="16dp"
                 android:clipToPadding="false"
+                android:overScrollMode="never"
                 android:paddingEnd="5dp"
                 android:paddingBottom="1dp"
                 app:layout_constraintEnd_toEndOf="parent"
